move shared cache md funcs to own pkg
